+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/InfoCommand.java
@@ -14,6 +14,7 @@ import io.github.scroojalix.npcmanager.utils.PluginUtils;
 import io.github.scroojalix.npcmanager.utils.interactions.NPCInteractionData;
 import io.github.scroojalix.npcmanager.utils.npc.NPCData;
 import io.github.scroojalix.npcmanager.utils.npc.NPCTrait;
+import io.github.scroojalix.npcmanager.utils.npc.skin.NPCSkinLayers;
 import io.github.scroojalix.npcmanager.utils.npc.skin.SkinData;
 
 public class InfoCommand extends SubCommand {
@@ -38,11 +39,10 @@ public class InfoCommand extends SubCommand {
         return true;
     }
 
-    //TODO show skin layers
     @Override
     public boolean execute(NPCMain main, CommandSender sender, String[] args) {
         if (args.length < 2)
-            return false;
+        return false;
         if (CommandUtils.npcExists(args[1], sender)) {
             NPCData data = main.npc.getNPCs().get(args[1]);
             NPCTrait traits = data.getTraits();
@@ -54,12 +54,7 @@ public class InfoCommand extends SubCommand {
                 sender.sendMessage(PluginUtils.format("&6Subtitle: &F" + traits.getSubtitle()));
                 sender.sendMessage(PluginUtils.format("&6Range: &F" + traits.getRange()));
                 sender.sendMessage(PluginUtils.format("&6Head Rotation: &F" + traits.hasHeadRotation()));
-                SkinData skinData = traits.getSkinData();
-                sender.sendMessage(PluginUtils.format("&6Skin: &F"+(skinData==null?null:skinData.getSkinName())));
                 CommandUtils.sendJSONMessage(sender, CommandUtils.getEquipmentComponents(data.getName()));
-                CommandUtils.sendJSONMessage(sender, CommandUtils.getPageTurnerMessage("/npc info " + data.getName(), 2, 1));
-            } else if (args[2].equalsIgnoreCase("2")) {
-                CommandUtils.sendJSONMessage(sender, CommandUtils.getTitleMessage("NPC Info"));
                 if (data.getInteractEvent() != null) {
                     NPCInteractionData interactEvent = traits.getInteractEvent();
                     sender.sendMessage(PluginUtils.format("&6Interact Event: "+interactEvent.getType().infoPrefix
@@ -67,9 +62,14 @@ public class InfoCommand extends SubCommand {
                 } else {
                     sender.sendMessage(PluginUtils.format("&6Interact Event: &fnull"));
                 }
+                CommandUtils.sendJSONMessage(sender, CommandUtils.getPageTurnerMessage("/npc info " + data.getName(), 2, 1));
+            } else if (args[2].equalsIgnoreCase("2")) {
+                CommandUtils.sendJSONMessage(sender, CommandUtils.getTitleMessage("NPC Info"));
+                SkinData skinData = traits.getSkinData();
+                sender.sendMessage(PluginUtils.format("&6Skin: &F"+(skinData==null?null:skinData.getSkinName())));
+                NPCSkinLayers layers = traits.getSkinLayers();
+                sender.sendMessage(ChatColor.GOLD+"Skin Layers: "+(layers==null?new NPCSkinLayers():layers).getCurrentConfiguration());
                 if (sender instanceof Player) {
-                    sender.sendMessage("");
-                    sender.sendMessage("");
                     sender.sendMessage("");
                     sender.sendMessage("");
                     sender.sendMessage("");